Move exmaples codes
