+++ Examples.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 7cf71f392af74beab6efd9f8b572de24
+timeCreated: 1647330058
