+++ Examples/1_UseLogger.cs
