+++ Examples/Edanoue.Logging.Examples.asmdef.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: e6398ed387334ae1872d06a0f794a136
+timeCreated: 1647330184
