--- Tests/Examples.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a77add31f96a9364e82e5eb4e6ac919b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
